+++ Libs/Allocator.h
@@ -1,8 +1,6 @@
 #ifndef Allocator_h__
 #define Allocator_h__
 
-#include <stdio.h>
-
-void* Allocator_Malloc(size_t size);
+void* Allocator_Malloc(int size);
 
 #endif // Allocator_h__
